requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
asyncpg
pandas
requests
python-dotenv
websocket-client
ta
tqdm
httpx
authlib
pydantic[email]
loguru
python-jose[cryptography]
celery
redis
aiohttp
aiolimiter